==> R/00_doit.R <==
# Run all scripts ---------------------------------------------------------
source(file = "R/98_check_packages.R")
source(file = "R/01_load.R")
source(file = "R/02_clean.R") #Runs slow, check README.txt
source(file = "R/03_augment.R")
source(file = "R/04_analysis_variable_exploration.R")
source(file = "R/05_analysis_diabetes_related_visualisations.R")
source(file = "R/06_analysis_PCA.R")
source(file = "R/07_analysis_k_means.R")
source(file = "R/08_analysis_Log_Reg.R")
